change
I made significant changes to complete the task according to the guidelines – instead of the login and register sections opening on the index page, they now have their own separate subpages. Because of that, I had to do some tweaking in the CSS, but for now everything looks as it should – ready for further styling. At this point, I’ll work on the register form to add hints and validations.